requirements.txt
discord.py>=2.3.0
slack-sdk>=3.21.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
asyncio-throttle>=1.0.0
aiosqlite>=0.19.0
openpyxl>=3.1.2